App Component
Getting help from GPT, these are the prompts I put in to learn more about why this setup may or may not be a great idea. There are definitely more improvements I can make here eventually, this is just the initial setup for things.

"What are the performance implications of having multiple render roots in a SolidJS project?"

"I want to write an abstraction layer on top of components, making use of Solid-JS's state management and syntax lifecycles. I will expand upon them to construct Web Components under the hood, and those are what will be resolved to when running in the final build itself."