Component Test
This is actually an awesome side-effect I'd never thought of before, since they are just functions either way, like SolidJS does, you can have a Web Component that doesn't *have* to be a Web Component if you don't need it to be! So if you already have a parent, you can just import the 'constructor' itself, and not worry about the parent scope. This is extra cool. Now I need to make sure that using `<num-text/>` will still work too, because right now in this context I don't think it will.